+++ tests/incrVersion/test_incrementVersion.py
@@ -21,7 +21,7 @@ class TestIncrementVersion (unittest.TestCase):
         for file in ret:
             self.assertTrue(doesFileExist(file))
         ret2 = incrementVersion._getJsonSchemaFileNames('./tests')
-        self.assertEqual(15, len(ret2))
+        self.assertEqual(17, len(ret2))
         for file in ret2:
             self.assertTrue(doesFileExist(file))
         # negative test