+++ yacg/model/modelFuncs.py
@@ -92,6 +92,20 @@ def flattenTypes(loadedTypes):
     return loadedTypes
 
 
+def processYacgTags(loadedTypes):
+    typesToReturn = []
+    for type in loadedTypes:
+        if isinstance(type, model.ComplexType):
+            if hasTag("yacgIgnoreForModel", type):
+                continue
+            if hasTag("yacgFlattenType", type):
+                flattenProperties = getFlattenProperties(type)
+                type.properties = flattenProperties
+                type.extendsType = None
+        typesToReturn.append(type)
+    return typesToReturn
+
+
 def isEnumType(typeObj):
     """checks if the given type object is an EnumType. If that's the
     case then True is returned, else the return is false